Mimics/AI-Resume-Analyzer/backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
spacy
google-generativeai